Use unsigned in for loop

+++ src/xmipp/libraries/reconstruction_cuda/cuda_forward_art_zernike3d.cpp
@@ -219,9 +219,9 @@ namespace {
 	std::tuple<unsigned *, size_t> filterAndTransportMask(MultidimArray<int> mask)
 	{
 		std::vector<unsigned> coordinates;
-		for (size_t i = 0; i < mask.yxdim * mask.zdim; i++) {
+		for (unsigned i = 0; i < mask.yxdim * mask.zdim; i++) {
 			if (mask[i] != 0) {
-				coordinates.push_back(static_cast<unsigned>(i));
+				coordinates.push_back(i);
 			}
 		}
 		return std::make_tuple(transportStdVectorToGpu(coordinates), coordinates.size());
